Revert "Engine and game are now in the same solution. Game.cpp has been removed from the game solution; it shouldn't have been there in the first place."
This reverts commit 1311dab3ef3049f9454a22ab1d02a86656e14535.

+++ source/Game/MapGenerator.cpp
@@ -104,7 +104,7 @@ void MapGenerator::createSectors()
 
 typeSector MapGenerator::getRandomType()
 {//(typeSector)(rand() % (TOTALTYPES - 2));
-	int i = rand() % 100;
+	float i = rand() % 100;
 	for(int j = 0; j < TOTALTYPES - 2; j++)
 	{
 		i -= typeChances[j];
@@ -234,13 +234,13 @@ bool MapGenerator::collisionLineBetweenSectors(MapSector* sector1, MapSector* se
 		if (quad >= 0)
 		{
 			// An infinite collision is happening, but let's not stop here
-			double quadsqrt = sqrt(quad);
+			float quadsqrt = sqrt(quad);
 			for (int i = -1; i <= 1; i += 2)
 			{
 				// Returns the two coordinates of the intersection points
-				double t = (i * -b + quadsqrt) / (2 * a);
-				double x = ax + (i * vx * t);
-				double y = ay + (i * vy * t);
+				float t = (i * -b + quadsqrt) / (2 * a);
+				float x = ax + (i * vx * t);
+				float y = ay + (i * vy * t);
 				// If one of them is in the boundaries of the segment, it collides
 				if (x >= min(ax, bx) && x <= max(ax, bx) && y >= min(ay, by) && y <= max(ay, by))
 				{
@@ -255,7 +255,7 @@ bool MapGenerator::collisionLineBetweenSectors(MapSector* sector1, MapSector* se
 std::string MapGenerator::nameGenerator(typeSector type)
 {
 	//Pick a random name and remove it from the list to avoid duplicates
-	int rnd = rand() % nameprefix.size();
+	float rnd = rand() % nameprefix.size();
 	std::string name = nameprefix.at(rnd) + " ";
 	nameprefix.erase(nameprefix.begin() + rnd);
